Refactor wiring in default.nix and nix/default.nix
Move sources, haskellNix and haskellNixOverlays to the top level
default.nix which allows for passing the haskellNixOverlays from
flake.nix and thus simply importing 'default.nix' directly.

The reason why this is important is that haskell.nix needs to use
a different overlay for when using in flake evaluationor normal one.
This is due to the fact that when evaluating flakes
builtins.currentSystem is not available.